+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Features:
   + Added "Constructor" FDEPL property that marks methods for being generated as
     constructors/initializers in Java/Swift.
+  + Added "Static" FDEPL property support for attributes.
 
 ## 3.18.0
 - Release date: 2018-12-18